pyproject.toml
[tool.poetry]
name = "binance-future2"
version = "0.1.0"
description = ""
authors = ["algorijun <123007169+HGive@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
python-binance = "^1.0.27"
pandas = "^2.2.3"
pandas-ta = "^0.3.14b0"
numpy = "^1.23.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
